Added Grade feature, incomplete

+++ src/main/java/com/f1soft/sces/repository/EnrollmentCourseRepository.java
@@ -3,6 +3,6 @@ package com.f1soft.sces.repository;
 import com.f1soft.sces.entities.EnrollmentCourse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnrollmentCourseRepository extends JpaRepository<EnrollmentCourse, Integer> {
+public interface EnrollmentCourseRepository extends JpaRepository<EnrollmentCourse, Long> {
 
 }
